ExerciseScenario renamed to Exercise. Changed the content of exercises to the old one.
